Second commit

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
         int amountTicket = 100;
-        int bonusAmount = 20;
-        int totalBonus = amountTicket / bonusAmount;
+        long bonusAmount = 20;
+        long totalBonus = amountTicket / bonusAmount;
         System.out.println(totalBonus);
     }
 }
